Decompile func00039e5c

+++ src/include/lib/lib_39c80.h
@@ -5,6 +5,6 @@
 #include "types.h"
 
 void func00039cd0(N_ALCSPlayer *seqp);
-void func00039e5c(N_ALCSPlayer *seqp, s32 arg1, s32 arg2, u8 arg3);
+void func00039e5c(N_ALCSPlayer *seqp, s32 arg1, u8 arg2, u8 arg3);
 
 #endif
